# openapi.yml
openapi: 3.0.0
info:
  title: Flow Wallet API
  description: A REST HTTP service that allows a developer to integrate wallet functionality into a larger Flow application infrastructure
  version: 0.9.0
  contact:
    url: "https://github.com/orgs/flow-hydraulics/teams/wallet"
servers:
  - url: "http://localhost:3000/v1"
...
paths:
...
  "/tokens/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: number
          example: 1
    delete:
      summary: Remove a token from the service
      description: "Remove a token, disabling functionality regarding it. Won't affect accounts."
      operationId: removeToken
      tags:
        - Fungible Tokens
        - Non-Fungible Tokens
      responses:
        "200":
          description: Removal OK.
          content:
            application/json:
              schema:
                type: integer
...